chore(imagetool): split provenance modules
